% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_doc.R
\name{ajusta_doc}
\alias{ajusta_doc}
\title{Ajusta Documentos no Dataframe}
\usage{
ajusta_doc(df)
}
\arguments{
\item{df}{Dataframe contendo as colunas de documentos a serem ajustadas.
Espera-se que pelo menos as colunas \code{nu_doc} e \code{nu_cns} estejam presentes no dataframe.}
}
\value{
Dataframe modificado com as colunas de documentos ajustadas.
As operações incluem a remoção de caracteres não numéricos, a aplicação de regras de substituição condicional,
a conversão de valores para \code{NA} quando apropriado, e a adição de novas colunas ajustadas \code{nu_doc_n}, \code{nu_cns_n},
e \code{nu_cpf} com valores possivelmente convertidos para formatos numéricos. A coluna \code{verificando} é adicionada
para indicar valores de documentos considerados válidos após o processo de verificação.
}
\description{
Esta função realiza várias operações de limpeza e ajuste em colunas específicas de documentos em um dataframe.
Inicialmente, remove espaços em branco no início e no fim dos valores nas colunas de documentos. Em seguida, identifica
e marca valores que contêm caracteres não numéricos. Esses valores são então substituídos por \code{NA}. A função também
aplica uma série de regras condicionais para substituir conteúdos específicos por \code{NA} com base em critérios pré-definidos
de validade ou relevância dos números de documentos. Por fim, a função tenta padronizar e corrigir os números de documentos,
convertendo-os para formatos numéricos quando possível e criando novas colunas para representar documentos verificados
e ajustados conforme as regras aplicadas.
}
