% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_encoding_char.R
\name{ajusta_encoding_char}
\alias{ajusta_encoding_char}
\title{Correção de encoding em colunas Character}
\usage{
ajusta_encoding_char(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna o encoding ajustado para todas as colunas no formato character no data frame
}
\description{
Correção de encoding em colunas Character
}
