% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_encoding_sim.R
\name{ajusta_encoding_sim}
\alias{ajusta_encoding_sim}
\title{Ajusta o encoding de NOMEPAI, NOME e NOMEMAE do SIM}
\usage{
ajusta_encoding_sim(df)
}
\arguments{
\item{df}{nome do data frame}
}
\value{
Retorna o encoding das colunas NOMEPAI, NOME e NOMEMAE do SIM corrigidos
com UTF-8, ASCII
}
\description{
Ajusta o encoding de NOMEPAI, NOME e NOMEMAE do SIM
}
