% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_cd_to_int.R
\name{ajuste_cd_to_int}
\alias{ajuste_cd_to_int}
\title{Ajusta colunas que começam com cd_ para números inteiros}
\usage{
ajuste_cd_to_int(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna as colunas que começam com "cd_" convertidas para números inteiros
}
\description{
Ajusta colunas que começam com cd_ para números inteiros
}
