% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_data.R
\name{ajuste_data}
\alias{ajuste_data}
\title{Ajuste nas colunas no formato de data}
\usage{
ajuste_data(df, tipo_data = 1)
}
\arguments{
\item{df}{Nome do data frame}

\item{tipo_data}{Tipo do formato de data: 1 para "YYYYMMDD" ou 2 para "YYYY-MM-DD"}
}
\value{
O dataframe modificado, onde as colunas que começam com "dt_" ou "DT_"
foram convertidas para o formato de data especificado pelo parâmetro tipo_data.
}
\description{
Ajuste nas colunas no formato de data
}
