% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_data_tipo1.R
\name{ajuste_data_tipo1}
\alias{ajuste_data_tipo1}
\title{Ajusta colunas que começam com dt_ no formato YYYYMMDD}
\usage{
ajuste_data_tipo1(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna as acolunas que começam com "dt_" corrigidas, desde que estejam no formato YYYYMMDD.
Exemplo: Se estiver 20101128 retornará como 2010-11-28
}
\description{
Ajusta colunas que começam com dt_ no formato YYYYMMDD
}
