% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_data_tipo2.R
\name{ajuste_data_tipo2}
\alias{ajuste_data_tipo2}
\title{Ajusta as colunas que começam com "dt_" no formato YYYY-MM-DD}
\usage{
ajuste_data_tipo2(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna as acolunas que começam com "dt_" corrigidas no formato de data, desde que estejam no formato YYYY-MM-DD.
Exemplo: Se estiver 2010-11-28 retornará como 2010-11-28 no formato correto.
}
\description{
Ajusta as colunas que começam com "dt_" no formato YYYY-MM-DD
}
