% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_doc.R
\name{ajuste_doc}
\alias{ajuste_doc}
\title{Ajusta colunas do tipo character que contêm "_doc" para números inteiros.}
\usage{
ajuste_doc(df)
}
\arguments{
\item{df}{data frame a ser ajustado.}
}
\value{
Um data frame modificado, onde as colunas do tipo character contendo "_doc" em seus
nomes foram transformadas em números inteiros, mantendo apenas os valores numéricos presentes.
}
\description{
Esta função realiza ajustes em um data frame, transformando as colunas do tipo character
que contêm "_doc" em seus nomes para números inteiros. Os ajustes incluem a remoção de
espaços extras, a transcrição de caracteres acentuados para ASCII e a remoção de caracteres
não numéricos.
}
