% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_res.R
\name{ajuste_res}
\alias{ajuste_res}
\title{Ajusta as colunas de texto do dataframe}
\usage{
ajuste_res(df)
}
\arguments{
\item{df}{Um dataframe a ser ajustado.}
}
\value{
Retorna o dataframe com os ajustes realizados.
}
\description{
Essa função realiza ajustes específicos nas colunas de texto de um dataframe, incluindo
a remoção de espaços extras, a transliteração de caracteres acentuados para ASCII,
a normalização para letras maiúsculas, e outras operações específicas para colunas
que contenham a substring "_res".
}
\details{
A função percorre cada coluna do dataframe e, se for uma coluna de texto e
contiver a substring "_res" (ignorando maiúsculas e minúsculas), realiza os seguintes ajustes:
\itemize{
\item Translitera caracteres acentuados para ASCII.
\item Converte o texto para letras maiúsculas.
\item Remove espaços extras.
\item Remove prefixos específicos, como "RECEM NASCIDO", "NATIMORTO", entre outros.
\item Remove sufixos específicos, como "FILHO", "NETO", "SOBRINHO", "JUNIOR", entre outros.
\item Normaliza a utilização de preposições ("DE", "DOS", "DA", "DAS", "DO", "DDAS", "E") para espaços.
\item Remove caracteres não alfabéticos e não numéricos.
\item Remove repetições excessivas de letras (mais de duas vezes).
\item Troca novamente 2 ou mais espaços por um único espaço.
}
}
\examples{
\dontrun{
  df <- data.frame(NOME_res = c("João da Silva", "Maria dos Santos", "José Neto Junior"))
  resultado <- ajuste_res(df)
}

}
