% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_txt.R
\name{ajuste_txt}
\alias{ajuste_txt}
\title{Ajusta colunas de texto em um dataframe}
\usage{
ajuste_txt(df)
}
\arguments{
\item{df}{Um dataframe a ser ajustado.}
}
\value{
Retorna o dataframe com os ajustes realizados.
}
\description{
Esta função realiza ajustes específicos nas colunas de texto de um dataframe, incluindo
a remoção de espaços extras, a transliteração de caracteres acentuados para ASCII,
a normalização para letras maiúsculas, e outras operações específicas para colunas
que contêm a substring "nome" (ignorando maiúsculas e minúsculas).
}
\details{
A função percorre cada coluna do dataframe e, se for uma coluna de texto e
contiver a substring "nome" (ignorando maiúsculas e minúsculas), realiza os seguintes ajustes:
\itemize{
\item Translitera caracteres acentuados para ASCII.
\item Converte o texto para letras maiúsculas.
\item Remove espaços extras.
\item Remove prefixos específicos, como "RECEM NASCIDO", "NATIMORTO", entre outros.
\item Remove sufixos específicos, como "FILHO", "NETO", "SOBRINHO", "JUNIOR", entre outros.
\item Normaliza a utilização de preposições ("DE", "DOS", "DA", "DAS", "DO", "DDAS", "E") para espaços.
\item Remove caracteres não alfabéticos e não numéricos.
\item Troca novamente 2 ou mais espaços por um único espaço.
\item Se a letra repete mais de duas vezes, remove uma.
}
}
\examples{
\dontrun{
  df <- data.frame(nome_completo = c("João da Silva", "Maria dos Santos", "José Neto Junior"))
  resultado <- ajuste_txt(df)
}

}
