% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ano_sih.R
\name{ano_sih}
\alias{ano_sih}
\title{Cria a coluna ANO a aprtir de dt_internacao}
\usage{
ano_sih(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna uma coluna nova chamada ANO
}
\description{
Cria a coluna ANO a aprtir de dt_internacao
}
