% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ano_sim.R
\name{ano_sim}
\alias{ano_sim}
\title{Cria a coluna ANO a aprtir de DTOBITO}
\usage{
ano_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna uma coluna nova chamada ANO
}
\description{
Cria a coluna ANO a aprtir de DTOBITO
}
