% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ano_sinan.R
\name{ano_sinan}
\alias{ano_sinan}
\title{Cria a coluna dt_nasc a aprtir da data de nascimento}
\usage{
ano_sinan(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna uma coluna nova chamada ANO
}
\description{
Cria a coluna dt_nasc a aprtir da data de nascimento
}
