% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_char.R
\name{as_char}
\alias{as_char}
\title{Transforma as variáveis em Character}
\usage{
as_char(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna o data frame com as variáveis no formato character
}
\description{
Transforma as variáveis em Character
}
