% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_data_comum.R
\name{atualiza_dt_comum}
\alias{atualiza_dt_comum}
\title{Atualiza Data Comum com Base em Condições Específicas}
\usage{
atualiza_dt_comum(df)
}
\arguments{
\item{df}{Dataframe que contém as colunas \code{dt_comum}, \code{dt_obito}, \code{dt_notific} e \code{banco}.}
}
\value{
Dataframe com a coluna \code{dt_comum} atualizada conforme as condições especificadas.
}
\description{
Esta função atualiza a coluna \code{dt_comum} de um dataframe com valores de outras colunas de datas (\code{dt_obito}, \code{dt_notific}),
baseando-se no valor da coluna \code{banco}. A atualização é feita conforme as seguintes regras:
\itemize{
\item Se \code{banco} for "SIM", \code{dt_comum} é atualizado para o valor de \code{dt_obito}.
\item Se \code{banco} for "SINAN", \code{dt_comum} é atualizado para o valor de \code{dt_notific}.
\item Se \code{banco} for "SESAP_OB", \code{dt_comum} é atualizado para o valor de \code{dt_obito}.
}
}
