% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_novos.R
\name{casos_novos}
\alias{casos_novos}
\title{Função para filtrar casos novos em um conjunto de dados}
\usage{
casos_novos(df, col)
}
\arguments{
\item{df}{Um data frame contendo os dados a serem filtrados.}

\item{col}{O nome da coluna que será utilizada para a filtragem de casos novos.}
}
\value{
Um data frame contendo apenas os casos novos conforme especificado pela coluna fornecida.
}
\description{
Esta função recebe um conjunto de dados e o nome de uma coluna para realizar
a filtragem de casos novos com base em uma variável criada previamente.
A função requer que a função \code{vitallinkage::start_linkage()} seja aplicada
anteriormente, pois ela cria a variável \code{par_1} necessária para o funcionamento
correto desta função.
}
