% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copia_nomes.R
\name{copia_nomes}
\alias{copia_nomes}
\title{Copia Coluna de Nomes}
\usage{
copia_nomes(df)
}
\arguments{
\item{df}{Um data frame contendo as variáveis a serem copiadas.}
}
\value{
Retorna o data frame com as novas colunas adicionadas.
}
\description{
Esta função copia os nomes originais das variáveis para novas colunas específicas,
apenas se as variáveis originais existirem no conjunto de dados.
}
