% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrige_datas.R
\name{corrige_datas}
\alias{corrige_datas}
\title{Corrige todas as colunas de data}
\usage{
corrige_datas(df)
}
\arguments{
\item{df}{nome do dataframe}
}
\value{
Transforma todas as colunas que comecem com "dt_" no formato data YYYY-MM-DD
}
\description{
Corrige todas as colunas de data
}
