% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrige_sg_sexo.R
\name{corrige_sg_sexo}
\alias{corrige_sg_sexo}
\title{Corretor de sg_sexo no SINAN viol}
\usage{
corrige_sg_sexo(df)
}
\arguments{
\item{df}{Nome do dataframe}
}
\value{
Retorna a correção da coluna "sg_sexo" do sinan
}
\description{
Corretor de sg_sexo no SINAN viol
}
