% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_coluna_id.R
\name{criar_coluna_id}
\alias{criar_coluna_id}
\title{Cria coluna com o id único para cada registro ('nome da base + 1:n)}
\usage{
criar_coluna_id(df, nome_base)
}
\arguments{
\item{df}{Nome do dataframe}

\item{nome_base}{Nome da base, exemplos: "SINAN", "SIM" ou "SIH"}
}
\value{
Ao aplicar a função, retornará uma nova coluna composta por "id" + nome da base indicada no parâmetro 'nome_base'
}
\description{
Cria coluna com o id único para cada registro ('nome da base + 1:n)
}
