% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dias_int_sih.R
\name{dias_int_sih}
\alias{dias_int_sih}
\title{Número de dias internado}
\usage{
dias_int_sih(df)
}
\arguments{
\item{df}{nome do data frame}
}
\value{
Retorna uma coluna "dias_int" com número de dias que a pessoa fico internada (dt_saida - dt_internacao)
}
\description{
Número de dias internado
}
