% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_duplicados_sim.R
\name{drop_duplicados_sim}
\alias{drop_duplicados_sim}
\title{Trata os duplicados do SIM}
\usage{
drop_duplicados_sim(df)
}
\arguments{
\item{df}{nome do data frame}
}
\value{
Retorna o dataframe com os duplicados removidos
}
\description{
Trata os duplicados do SIM
}
