% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_duplicados_sim_padronizado.R
\name{drop_duplicados_sim_padronizado}
\alias{drop_duplicados_sim_padronizado}
\title{Remove duplicatas específicas do dataframe.}
\usage{
drop_duplicados_sim_padronizado(df)
}
\arguments{
\item{df}{Um dataframe contendo os dados a serem processados.}
}
\value{
Um dataframe sem as duplicatas identificadas pelas regras estabelecidas.
}
\description{
Esta função remove duplicatas do dataframe de acordo com regras específicas.
}
