% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_duplicados_sinan_1.R
\name{drop_duplicados_sinan_1}
\alias{drop_duplicados_sinan_1}
\title{Dropa registros duplicados no SINAN}
\usage{
drop_duplicados_sinan_1(df)
}
\arguments{
\item{df}{Nome do dataframe}
}
\value{
A função considera duas regras de duplicados:
Inicialmente ele verifica se existem registros exatamente iguais, se sim, aplica o drop deixando apenas um registro;
Em seguida, as colunas NU_NOTIFIC e NM_PACIENT são concatenadas. Se houverem duplicadas nessas regras, um dos registros será deletado.
}
\description{
Dropa registros duplicados no SINAN
}
