% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_raca_sih.R
\name{ds_raca_sih}
\alias{ds_raca_sih}
\title{Decodifica raça/cor do SIH}
\usage{
ds_raca_sih(df)
}
\arguments{
\item{df}{Nome do dataframe do SIH}
}
\value{
Nova coluna chamada "ds_raca" com a descrição da raça/cor
}
\description{
Decodifica raça/cor do SIH
}
