% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_raca_sim.R
\name{ds_raca_sim}
\alias{ds_raca_sim}
\title{Decodifica raça/cor do SIM}
\usage{
ds_raca_sim(df)
}
\arguments{
\item{df}{Nome do dataframe do SIM}
}
\value{
Nova coluna chamada "ds_raca" com a descrição da raça/cor
}
\description{
Decodifica raça/cor do SIM
}
