% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_raca_sinan.R
\name{ds_raca_sinan}
\alias{ds_raca_sinan}
\title{Decodifica raça/cor}
\usage{
ds_raca_sinan(df)
}
\arguments{
\item{df}{Nome do dataframe do SINAN}
}
\value{
Nova coluna chamada "ds_raca" com a descrição da raça/cor
}
\description{
Decodifica raça/cor
}
