% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro_par_c_especifico.R
\name{filtro_par_c_especifico}
\alias{filtro_par_c_especifico}
\title{Filter a DataFrame for a Specific 'par_c' Column}
\usage{
filtro_par_c_especifico(df, coluna_especifica)
}
\arguments{
\item{df}{A dataframe containing the data to be filtered.}

\item{coluna_especifica}{The specific 'par_c' column name to be checked for non-NA values.}
}
\value{
A dataframe filtered based on the specified conditions. The function first identifies all
columns starting with "par_c" and dynamically constructs a filter expression. It filters the dataframe
to include only rows where the specified 'par_c' column has non-NA values, and all other 'par_c' columns
have NA values. Additionally, it further filters rows where 'par_1' does not match the value of the
specified 'par_c' column.
}
\description{
This function filters a given dataframe based on the condition that only one specific 'par_c' column
should have non-NA values, while all other 'par_c' columns should have NA values. It further filters
rows based on a comparison between 'par_1' and the specific 'par_c' column's values.
}
