% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemelar.R
\name{gemelar}
\alias{gemelar}
\title{Identifica e Marca Pacientes com Dados Gemelares}
\usage{
gemelar(df, coluna)
}
\arguments{
\item{df}{Dataframe contendo os dados dos pacientes.}

\item{coluna}{Nome da coluna no dataframe que contém os nomes ou descrições dos pacientes a serem verificados.}
}
\value{
Dataframe original com uma coluna adicional 'gemelar', onde cada linha
que corresponde a um registro gemelar é marcada com 1, e 0 caso contrário.
}
\description{
Esta função percorre um dataframe e marca as linhas que contêm indicações
de dados gemelares na coluna especificada. Utiliza expressões regulares para
identificar várias indicações de gemelaridade, como "GEMELAR", "GEMEO", "GEMEA",
entre outros padrões. Uma nova coluna chamada 'gemelar' é criada (ou atualizada,
se já existir), onde cada linha correspondente a um paciente gemelar é marcada com 1.
}
