% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idade_sih.R
\name{idade_sih}
\alias{idade_sih}
\title{Cria a coluna de idade do SIH, quando preciso}
\usage{
idade_sih(df)
}
\arguments{
\item{df}{nome do Data frame}
}
\value{
Retorna a coluna cd_idade como número inteiro
}
\description{
Cria a coluna de idade do SIH, quando preciso
}
