% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa_ignorados_sim.R
\name{limpa_ignorados_sim}
\alias{limpa_ignorados_sim}
\title{Limpa textos sobre ignoados}
\usage{
limpa_ignorados_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna NA nos valores das colunas que contenham textos como "ignorados", etc
}
\description{
Limpa textos sobre ignoados
}
