% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meio_de_campo.R
\name{meio_de_campo}
\alias{meio_de_campo}
\title{Função meio_de_campo}
\usage{
meio_de_campo(df)
}
\arguments{
\item{df}{Um data frame contendo os dados de entrada.}
}
\value{
Um data frame modificado com ajustes feitos de acordo com as condições especificadas.
}
\description{
Esta função manipula um quadro de dados com base em condições e operações específicas.
}
