% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_idade_anos_sih.R
\name{nu_idade_anos_sih}
\alias{nu_idade_anos_sih}
\title{Croluna da idade em anos}
\usage{
nu_idade_anos_sih(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna a coluna nu_idade_anos
}
\description{
Croluna da idade em anos
}
