% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_idade_anos_sinan.R
\name{nu_idade_anos_sinan}
\alias{nu_idade_anos_sinan}
\title{Cria as idades do sinan}
\usage{
nu_idade_anos_sinan(df)
}
\arguments{
\item{df}{Nome do dataframe}
}
\value{
Retorna a coluna "nu_idade_anos" com as idades nos registros
}
\description{
Cria as idades do sinan
}
