% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_SIH.R
\name{padroniza_SIH}
\alias{padroniza_SIH}
\title{Tratamento da base SIH}
\usage{
padroniza_SIH(df, tipo_de_data = 2)
}
\arguments{
\item{df}{Data frame com os dados do SIH}

\item{tipo_de_data}{O formato de data que deve retornar}
}
\value{
Base com os dados do SIH tratados
}
\description{
Tratamento da base SIH
}
