% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_SINAN.R
\name{padroniza_SINAN}
\alias{padroniza_SINAN}
\title{Padroniza e realiza ajustes nos dados do SINAN.}
\usage{
padroniza_SINAN(df, tipo_de_data = 2)
}
\arguments{
\item{df}{DataFrame contendo os dados do SINAN.}

\item{tipo_de_data}{(opcional) Tipo de ajuste para o formato de data. Pode ser 1 para formato DD/MM/AAAA ou 2 para AAAA-MM-DD.}
}
\value{
DataFrame contendo os dados do SINAN após as transformações.
}
\description{
Esta função padroniza e realiza ajustes nos dados do Sistema de Informação de Agravos de Notificação (SINAN)
}
