% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regras_linkage.R
\name{regras_linkage}
\alias{regras_linkage}
\title{Aplica Regras de Linkage em um Dataframe}
\usage{
regras_linkage(df, variables, chave, num_regra)
}
\arguments{
\item{df}{Dataframe sobre o qual as regras de linkage serão aplicadas.}

\item{variables}{Vetor de caracteres contendo os nomes das colunas usadas para organizar e agrupar os dados.}

\item{chave}{Vetor de caracteres contendo os nomes das colunas usadas como chaves para identificar registros relacionados.}

\item{num_regra}{Número inteiro que identifica a regra de linkage específica a ser aplicada.}
}
\value{
Dataframe modificado após a aplicação das regras de linkage. O dataframe resultante inclui colunas adicionais
indicando a aplicação das regras e a identificação de grupos relacionados.
}
\description{
Esta função aplica regras de linkage específicas em um dataframe, organizando e agrupando dados
com base em variáveis específicas e critérios de chave. Ela permite a aplicação dinâmica de regras
de linkage, facilitando a identificação de registros correspondentes ou relacionados.
}
