% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sih_anon.R
\name{sih_anon}
\alias{sih_anon}
\title{Anonimiza SIH}
\usage{
sih_anon(df)
}
\arguments{
\item{df}{nome do data frame}
}
\value{
Retorna o data frame anonimizado, removendo as colunas com informações pessoais
}
\description{
Anonimiza SIH
}
