% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_anon.R
\name{sim_anon}
\alias{sim_anon}
\title{Anonimiza o dataframe do SIM}
\usage{
sim_anon(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna o dataframe sem as colunas com caracteres identificaveis
}
\description{
Anonimiza o dataframe do SIM
}
