% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinan_anon.R
\name{sinan_anon}
\alias{sinan_anon}
\title{Anonimiza o dataframe do SINAN}
\usage{
sinan_anon(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna o dataframe sem as colunas com caracteres identificaveis
}
\description{
Anonimiza o dataframe do SINAN
}
