% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex_linkage.R
\name{soundex_linkage}
\alias{soundex_linkage}
\title{Generaliza a separação de componentes de um nome e aplica SoundexBR}
\usage{
soundex_linkage(df, col_name)
}
\arguments{
\item{df}{Um data frame contendo a coluna de nomes.}

\item{col_name}{O nome da coluna que contém os nomes completos.}
}
\value{
O data frame de entrada com colunas adicionais para as partes do nome e os códigos SoundexBR.
}
\description{
Esta função aceita um data frame e o nome de uma coluna contendo nomes completos.
Ela realiza a separação dos componentes do nome e aplica o algoritmo SoundexBR
para cada parte do nome, gerando novas colunas com os resultados.
}
\examples{
\dontrun{
  df <- data.frame(NOME = c("John Doe", "Jane Smith"))
  resultado <- soundex_linkage(df, "NOME")
}

}
