% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex_nome_sim.R
\name{soundex_nome_sim}
\alias{soundex_nome_sim}
\title{Aplica SoundexBR na coluna de NOME do SIM}
\usage{
soundex_nome_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna colunas com os nomes transformados em sons para o computador
}
\description{
Aplica SoundexBR na coluna de NOME do SIM
}
