% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex_nomemae_sim.R
\name{soundex_nomemae_sim}
\alias{soundex_nomemae_sim}
\title{Aplica SoundexBR na coluna de NOMEMAE do SIM}
\usage{
soundex_nomemae_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna colunas com os nomes da mãe transformado em sons para o computador
}
\description{
Aplica SoundexBR na coluna de NOMEMAE do SIM
}
