% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex_nomepai_sim.R
\name{soundex_nomepai_sim}
\alias{soundex_nomepai_sim}
\title{Aplica SoundexBR na coluna de NOMEPAI do SIM}
\usage{
soundex_nomepai_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Retorna colunas com os nomes transformados em sons para o computador
}
\description{
Aplica SoundexBR na coluna de NOMEPAI do SIM
}
