% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_linkage.R
\name{start_linkage}
\alias{start_linkage}
\title{Iniciar o processo de ligação de registros em um dataframe}
\usage{
start_linkage(df, variables, chave)
}
\arguments{
\item{df}{Um dataframe contendo os dados que serão utilizados no processo de ligação de registros.}

\item{variables}{Um vetor de caracteres contendo os nomes das variáveis a serem usadas no processo de ligação de registros.}

\item{chave}{Uma string contendo o nome da variável que servirá como chave única para identificar registros semelhantes.}
}
\value{
Retorna um dataframe com colunas adicionais contendo informações sobre a ligação de registros, como pares de registros semelhantes e identificadores de grupo.
}
\description{
Esta função realiza as etapas iniciais do processo de ligação de registros em um dataframe, incluindo ordenação, agrupamento, cálculo de pares de registros semelhantes e aplicação de regras de ligação.
}
