% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratamentos_txt_sim.R
\name{tratamentos_txt_sim}
\alias{tratamentos_txt_sim}
\title{Tratamento textual no SIM}
\usage{
tratamentos_txt_sim(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Colunas com os valores textuais tratados e padronizados
}
\description{
Tratamento textual no SIM
}
