% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_case_char.R
\name{upper_case_char}
\alias{upper_case_char}
\title{Upper nos textos das colunas de Character}
\usage{
upper_case_char(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Todas as colunas que estiverem como Character retornarão com texto em caixa alta
}
\description{
Upper nos textos das colunas de Character
}
