% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_case_ds.R
\name{upper_case_ds}
\alias{upper_case_ds}
\title{Upper nos textos das colunas de Character}
\usage{
upper_case_ds(df)
}
\arguments{
\item{df}{Nome do data frame}
}
\value{
Todas as colunas que começam com "ds_" e estiverem como Character retornarão com texto em caixa alta
}
\description{
Upper nos textos das colunas de Character
}
