% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variaveis_principais_sim.R
\name{variaveis_principais_sim}
\alias{variaveis_principais_sim}
\title{Variáveis principais do SIM}
\usage{
variaveis_principais_sim(df)
}
\arguments{
\item{df}{nome do Data frame}
}
\value{
Data frame reduzido, apenas com as colunas necessárias
}
\description{
Variáveis principais do SIM
}
